+++ R/akin-defunct.R
@@ -1,0 +1,10 @@
+#'@title Defunct Functions In Package akin
+#'
+#'@description The following functions are defunct. Where available, alternative functions are suggested
+#' at \code{help("akin-defunct")}.
+#'
+#' @section \code{common}: this function is defunct. Use [fcommon] instead.
+#' @keywords internal
+#' @name akin-defunct
+#'
+NULL