+++ R/zzz.R
@@ -1,24 +1,18 @@
 
 #' @import data.table
 #' @import RVerbalExpressions
-#' @import RcppAlgos
 #' @import methods
+#' @importFrom RcppAlgos comboGeneral comboCount
 #' @importFrom utils type.convert getFromNamespace
 #' @importFrom fastmatch fmatch
 #' @importFrom stats median setNames
 #' @importFrom callr r_bg
 #' @importFrom Matrix Matrix rowSums rsparsematrix
 #' @importFrom listenv listenv
-#' @importFrom future.apply future_lapply
 #'
 
      .onLoad = function(libname, pkgname) {
                          Mmatrix = getFromNamespace('Mmatrix', 'Matrix')
                    rsparsematrix = getFromNamespace('rsparsematrix', 'Matrix')
 
-                  }
-   .onAttach = function(libname, pkgname) {
-                      suppressMessages(
-                                   requireNamespace("future", quietly = TRUE)
-                                )
                   }